# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agile Project Management API. Manages Projects → User Stories → Tasks for small teams, with asynchronous email notifications.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects with summary counts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with its stories and tasks
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (cascades to stories and tasks)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/stories:
    get:
      operationId: listStories
      tags: [stories]
      summary: List user stories for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StorySummary"
    post:
      operationId: createStory
      tags: [stories]
      summary: Create a user story under a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{storyId}:
    get:
      operationId: getStory
      tags: [stories]
      summary: Get a single user story with tasks
      parameters:
        - $ref: "#/components/parameters/StoryId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateStory
      tags: [stories]
      summary: Update a user story
      parameters:
        - $ref: "#/components/parameters/StoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoryInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
    delete:
      operationId: deleteStory
      tags: [stories]
      summary: Delete a user story (cascades to tasks)
      parameters:
        - $ref: "#/components/parameters/StoryId"
      responses:
        "204":
          description: Deleted
...
  /notifications/{notificationId}/retry:
    post:
      operationId: retryNotification
      tags: [notifications]
      summary: Manually requeue a failed notification
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
    StoryId:
      name: storyId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Member:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
...
    Project:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
          description: Short uppercase key (e.g. "WEB", "API")
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived, completed]
        ownerId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, key, name, description, status, ownerId, createdAt, updatedAt]
    ProjectSummary:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            owner:
              oneOf:
                - $ref: "#/components/schemas/Member"
                - type: "null"
            storyCount:
              type: integer
            taskCount:
              type: integer
            tasksDone:
              type: integer
            tasksInProgress:
              type: integer
            tasksTodo:
              type: integer
          required: [owner, storyCount, taskCount, tasksDone, tasksInProgress, tasksTodo]
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          properties:
            stories:
              type: array
              items:
                $ref: "#/components/schemas/StoryDetail"
          required: [stories]
    CreateProjectInput:
      type: object
      properties:
        key:
          type: string
          minLength: 1
          maxLength: 10
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
      required: [key, name]
    UpdateProjectInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived, completed]
        ownerId:
          type: ["integer", "null"]
    Story:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        acceptanceCriteria:
          type: ["string", "null"]
        status:
          type: string
          enum: [backlog, ready, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        storyPoints:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, title, description, acceptanceCriteria, status, priority, storyPoints, createdAt, updatedAt]
    StorySummary:
      allOf:
        - $ref: "#/components/schemas/Story"
        - type: object
          properties:
            taskCount:
              type: integer
            tasksDone:
              type: integer
          required: [taskCount, tasksDone]
    StoryDetail:
      allOf:
        - $ref: "#/components/schemas/Story"
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
          required: [tasks]
    CreateStoryInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        acceptanceCriteria:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        storyPoints:
          type: ["integer", "null"]
      required: [title]
    UpdateStoryInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        acceptanceCriteria:
          type: ["string", "null"]
        status:
          type: string
          enum: [backlog, ready, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        storyPoints:
          type: ["integer", "null"]
    Task:
      type: object
      properties:
        id:
          type: integer
        storyId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, blocked, done]
        assigneeId:
          type: ["integer", "null"]
        assignee:
          oneOf:
            - $ref: "#/components/schemas/Member"
            - type: "null"
        estimateHours:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, storyId, title, description, status, assigneeId, assignee, estimateHours, dueDate, createdAt, updatedAt]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        kind:
          type: string
          description: "Notification kind (e.g. task.assigned, task.status_changed)"
        recipientEmail:
          type: string
        recipientName:
          type: ["string", "null"]
        subject:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [pending, sending, sent, failed, dead_letter]
        attempts:
          type: integer
        maxAttempts:
          type: integer
        lastError:
          type: ["string", "null"]
        nextAttemptAt:
          type: ["string", "null"]
          format: date-time
        sentAt:
          type: ["string", "null"]
          format: date-time
        relatedTaskId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, kind, recipientEmail, recipientName, subject, body, status, attempts, maxAttempts, lastError, nextAttemptAt, sentAt, relatedTaskId, createdAt, updatedAt]